bug修复

+++ hjblogs-model-admin/src/main/java/cn/hjblogs/hjblogs/admin/service/impl/AdminCategoryServiceImpl.java
@@ -6,7 +6,9 @@ import cn.hjblogs.hjblogs.admin.model.vo.category.DeleteCategoryReqVO;
 import cn.hjblogs.hjblogs.admin.model.vo.category.FindCategoryPageListReqVO;
 import cn.hjblogs.hjblogs.admin.model.vo.category.FindCategoryPageListRspVO;
 import cn.hjblogs.hjblogs.admin.service.AdminCategoryService;
+import cn.hjblogs.hjblogs.common.domain.dos.ArticleCategoryRelDO;
 import cn.hjblogs.hjblogs.common.domain.dos.CategoryDO;
+import cn.hjblogs.hjblogs.common.domain.mapper.ArticleCategoryRelMapper;
 import cn.hjblogs.hjblogs.common.domain.mapper.CategoryMapper;
 import cn.hjblogs.hjblogs.common.enums.ResponseCodeEnum;
 import cn.hjblogs.hjblogs.common.exception.BizException;
@@ -36,6 +38,10 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private ArticleCategoryRelMapper articleCategoryRelMapper;
+
 
     /**
      * 添加分类
@@ -112,8 +118,24 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
 
     @Override
     public Response deleteCategory(DeleteCategoryReqVO deleteCategoryReqVO) {
+        /*// 分类 ID
+        Long categoryId = deleteCategoryReqVO.getId();
+
+        // 删除分类
+        categoryMapper.deleteById(categoryId);
+
+        return Response.success();*/
+
         // 分类 ID
         Long categoryId = deleteCategoryReqVO.getId();
+
+        // 校验该分类下是否已经有文章，若有，则提示需要先删除分类下所有文章，才能删除
+        ArticleCategoryRelDO articleCategoryRelDO = articleCategoryRelMapper.selectOneByCategoryId(categoryId);
+
+        if (Objects.nonNull(articleCategoryRelDO)) {
+            log.warn("==> 此分类下包含文章，无法删除，categoryId: {}", categoryId);
+            throw new BizException(ResponseCodeEnum.CATEGORY_CAN_NOT_DELETE);
+        }
 
         // 删除分类
         categoryMapper.deleteById(categoryId);
